Refactor container io code
