+++ pkg/server/container_create.go
@@ -175,10 +175,7 @@ func (c *criContainerdService) CreateContainer(ctx context.Context, r *runtime.C
 	}
 
 	containerIO, err := cio.NewContainerIO(id,
-		cio.WithStdinOpen(config.GetStdin()),
-		cio.WithTerminal(config.GetTty()),
-		cio.WithRootDir(containerRootDir),
-	)
+		cio.WithNewFIFOs(containerRootDir, config.GetTty(), config.GetStdin()))
 	if err != nil {
 		return nil, fmt.Errorf("failed to create container io: %v", err)
 	}
