+++ pkg/server/container_start.go
@@ -121,9 +121,7 @@ func (c *criContainerdService) startContainer(ctx context.Context,
 		if err := cio.WithOutput("log", stdoutWC, stderrWC)(cntr.IO); err != nil {
 			return nil, fmt.Errorf("failed to add container log: %v", err)
 		}
-		if err := cntr.IO.Pipe(); err != nil {
-			return nil, fmt.Errorf("failed to pipe container io: %v", err)
-		}
+		cntr.IO.Pipe()
 		return cntr.IO, nil
 	}
